chatbot/requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
vkbottle
loguru
aiogram
